# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portal Berita",
        "description": "Portal Berita API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Local server"
        },
        {
            "url": "http://staging.example.com",
            "description": "Staging server"
        },
        {
            "url": "http://example.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/buku/kategori/{id}": {
            "get": {
                "tags": [
                    "Buku"
                ],
                "summary": "Menampilkan buku berdasarkan ID kategori",
                "operationId": "bukuByKategori",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kategori buku",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data buku berdasarkan kategori",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil mengambil data buku berdasarkan kategori"
                                        },
                                        "kategori": {
                                            "$ref": "#/components/schemas/KategoriBukuSchema"
                                        },
                                        "buku": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BukuSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kategori tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori buku tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Kesalahan server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan saat mengambil data buku"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/kategori": {
            "get": {
                "tags": [
                    "KategoriBuku"
                ],
                "summary": "Menampilkan semua kategori",
                "operationId": "getKategoriBukuList",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data kategori",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int64",
                                                        "example": 1
                                                    },
                                                    "nama_kategori": {
                                                        "type": "string",
                                                        "example": "Fiksi"
                                                    },
                                                    "deskripsi": {
                                                        "type": "string",
                                                        "example": "Buku-buku fiksi dan novel"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-01-01T00:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-01-01T00:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Kesalahan server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan saat mengambil data kategori"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "KategoriBuku"
                ],
                "summary": "Menambahkan kategori baru",
                "operationId": "createKategoriBuku",
                "requestBody": {
                    "description": "Data kategori yang akan ditambahkan",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_kategori"
                                ],
                                "properties": {
                                    "nama_kategori": {
                                        "type": "string",
                                        "example": "Novel"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Kategori untuk buku novel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategori berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori berhasil ditambahkan"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "format": "int64",
                                                    "example": 1
                                                },
                                                "nama_kategori": {
                                                    "type": "string",
                                                    "example": "Novel"
                                                },
                                                "deskripsi": {
                                                    "type": "string",
                                                    "example": "Kategori untuk buku novel"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-01-01T00:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-01-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "nama_kategori": [
                                                    "Nama kategori harus diisi"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Kesalahan server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "KategoriBukuSchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_kategori": {
                        "type": "string",
                        "example": "Fiksi"
                    },
                    "deskripsi": {
                        "type": "string",
                        "example": "Buku-buku fiksi dan novel"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "BukuSchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "kategori_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_buku": {
                        "type": "string",
                        "example": "Harry Potter"
                    },
                    "judul": {
                        "type": "string",
                        "example": "Harry Potter and the Philosopher's Stone"
                    },
                    "penulis": {
                        "type": "string",
                        "example": "J.K. Rowling"
                    },
                    "penerbit": {
                        "type": "string",
                        "example": "Gramedia"
                    },
                    "tahun_penerbitan": {
                        "type": "string",
                        "format": "date",
                        "example": "2001-01-01"
                    },
                    "jumlah_tersedia": {
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "kategori": {
                        "$ref": "#/components/schemas/KategoriBukuSchema"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}